Move plotting files into dedicated subdirectory
